x86_64: page fault error codes

+++ src/platform/x86_64/include/platform/mmap.h
@@ -11,5 +11,5 @@
 #define PAGE_SIZE               4096            // bytes
 #define IDENTITY_MAP_GBS        16              // how many GiB of memory will be identity mapped
 #define KERNEL_BASE_ADDRESS     0x200000        // 2 MiB
-#define KERNEL_HEAP_BASE        (IDENTITY_MAP_GBS << 30)    // 16 GB
-#define USER_BASE_ADDRESS       (IDENTITY_MAP_GBS << 31)    // 32 GB
+#define KERNEL_HEAP_BASE        ((uint64_t)IDENTITY_MAP_GBS << 30)    // 16 GB
+#define USER_BASE_ADDRESS       ((uint64_t)IDENTITY_MAP_GBS << 31)    // 32 GB
